+++ src/api.rs
@@ -24,7 +24,9 @@ use std::fmt;
 
 use crate::error::Error;
 use crate::error::Error::{ClientErrorResponse, NotFound, ServerErrorResponse};
-use crate::requests::{Amqp091ShovelParams, EmptyPayload, StreamParams, SHOVEL_COMPONENT};
+use crate::requests::{
+    Amqp091ShovelParams, Amqp10ShovelParams, EmptyPayload, StreamParams, SHOVEL_COMPONENT,
+};
 use crate::responses::{
     DeprecatedFeatureList, FeatureFlag, FeatureFlagList, FeatureFlagStability, FeatureFlagState,
     GetMessage, OAuthConfiguration, SchemaDefinitionSyncStatus, VirtualHostDefinitionSet,
@@ -1374,6 +1376,25 @@ where
     }
 
     pub async fn declare_amqp091_shovel(&self, params: Amqp091ShovelParams<'_>) -> Result<()> {
+        let runtime_param = RuntimeParameterDefinition::from(params);
+
+        let _response = self
+            .http_put(
+                path!(
+                    "parameters",
+                    SHOVEL_COMPONENT,
+                    runtime_param.vhost,
+                    runtime_param.name
+                ),
+                &runtime_param,
+                None,
+                None,
+            )
+            .await?;
+        Ok(())
+    }
+
+    pub async fn declare_amqp10_shovel(&self, params: Amqp10ShovelParams<'_>) -> Result<()> {
         let runtime_param = RuntimeParameterDefinition::from(params);
 
         let _response = self